requirements.txt
Django~=5.1.4
psycopg2-binary==2.9.10
djangorestframework~=3.15.2
requests~=2.32.3